+++ src/navigation/DrawerNavigator.js
@@ -1,0 +1,20 @@
+import React, { useEffect } from 'react';
+import { createDrawerNavigator } from '@react-navigation/drawer';
+import Home from '../screens/Home';
+import * as ScreenOrientation from 'expo-screen-orientation';
+
+const Drawer = createDrawerNavigator();
+
+const DrawerNavigator = () => {
+  useEffect(() => {
+    ScreenOrientation.lockAsync(ScreenOrientation.OrientationLock.LANDSCAPE_LEFT);
+  }, []);
+
+  return (
+    <Drawer.Navigator initialRouteName="Home">
+      <Drawer.Screen name="Home" component={Home} />
+    </Drawer.Navigator>
+  );
+};
+
+export default DrawerNavigator;
